refactor: extract writeFileWithDirs + rename mergedHasBlocking (audit/tech-debt #2-3)
Two focused refactors from the v0.10.0 audit/tech-debt.md burn-down.
No user-visible behavior change in either.

    cmd/local-review/audit.go:writeAuditFile,
    bench.go:writeBenchToFile, bench.go:writeSWEBenchToFile all
    implemented the same mkdir + create + deferred-close-with-error-
    check sequence. New helper takes (path, emit func(io.Writer) error)
    and owns that contract; the three call sites become one-line
    wrappers that close over their report value. File mode (0644),
    directory mode (0o755), and close-error-doesn't-shadow-emit-error
    precedence are preserved exactly.

    Same function, clearer name. The audit flagged runner.go:478 as
    duplication; that claim was a false positive (there's only one
    implementation) but the function name was misleading — it's
    called both against the merged report AND against each per-LLM
    Output (via anyPerLLMHasBlocking as the truncation-safe backstop).
    The rename makes the symmetry visible at the call site. Doc
    comment now spells out both call contexts.
