+++ CHANGELOG.md
@@ -15,6 +15,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - **Dead code + dead parameters surfaced by v0.10.0's `audit/tech-debt.md`.** Three small deletions, no user-visible behavior change, opens the v0.10.2 burn-down on the audit's outstanding findings. (1) `internal/review/review.go` — removed the unused `matchGlob(path, pattern string) bool` function. It was lower-case (private), claimed in its doc comment to be "kept for back-compat with any external caller," but private functions have no external callers; production filtering goes through `compileGlobs` + `matchesAnyCompiled` (amortises regex compilation across the diff), tests go through `matchesAny`. Audit flagged it as dead code on `review.go:306-315`; the audit's framing was right. (2) `internal/cli/invoker.go` — removed the unused `errLabel string` parameter from `ClaudeInvoker.run` and its two callers (`Review`, `RunPrompt`). Pre-fix the parameter was passed in by every caller and explicitly discarded via `_ = errLabel` — a vestige of a pre-v0.7 "claude review failed:" prefix that the runner's per-LLM completion line now owns. CodexInvoker's `runExec` retains its `errLabel` parameter because it still distinguishes pre-invocation errors (temp-file creation) from post-success errors (output-file read), neither of which the runner's per-LLM line covers. (3) `internal/cli/version.go` — removed the unused `name` parameter from `detectVersion(name, path string)`. The function never read `name`; the audit's warning that "parameter suggests the function is name-aware when it's not" was accurate. Doc comment now points future maintainers at the right discriminator (`path`'s basename) if a CLI ever needs version-flag-specific branching.
 
+### Changed
+
+- **File-write boilerplate consolidated into `writeFileWithDirs` (audit/tech-debt #2).** `cmd/local-review/audit.go:writeAuditFile`, `bench.go:writeBenchToFile`, and `bench.go:writeSWEBenchToFile` all implemented the same `mkdir + create + deferred-close-with-error-check` sequence. New helper in `cmd/local-review/iohelpers.go` takes `(path string, emit func(io.Writer) error)` and owns that contract; the three call sites become one-line wrappers that close over their report value. No user-visible behavior change — file mode (0644), directory mode (0o755), close-error-doesn't-shadow-emit-error precedence are all preserved exactly.
+
+- **`mergedHasBlocking` → `isBlockingMarkdown` (audit/tech-debt #3).** Same function, clearer name. The audit flagged `runner.go:478` as duplication; the duplication claim was a false positive (there's only one implementation), but the function name was misleading enough that the auditor couldn't tell — it's called both against the merged report AND against each per-LLM Output (via `anyPerLLMHasBlocking` as the truncation-safe backstop). The rename makes the symmetry visible at the call site without changing behavior. Doc comment now spells out both call contexts so the next reader doesn't have to grep.
+
 ## [0.10.1] - 2026-05-25
 
 **Theme: fail fast on the slow path.**
